Add files via upload
This file can process the targets into four different datasets that can be used for classification:
Regular: A Basic 0/1 Classification for Alzheimer's
Pre-Clinical: This is classified as 1 if Alzheimer's is detected and the age of the patient is less than 60. 
ADMCI: Mild-Cognitive Impairment due to AD. 1 if AD is 1 and MMSE score is greater than 26. 
MMSE: The patients are classified based on their MMSE scores. This has more than 2 categories based on ranges of values. 0 for MMSE > 24, 1 for MMSE between 20 and 24, 2 for MMSE between 13 and 24 and 3 for MMSE anything less than 13. 
0: No Dementia
1: Mild Dementia
2: Moderate Dementia
3: Severe Dementia