Add files via upload
There are 8 datasets here corresponding to different targets:
2 files each, (in .csv format) for:
Regular diagnosis of 0/1 of AD
ADMCI - Dataset for Alzheimer's and Mild Cognitive Impairment, again modelled as 0/1 for the targets
PreClinical - Relationship between Diagnosis and Age. Again modelled as 0/1 for the targets
MMSE - MMSE values are a good indicator of the severity of Alzheimer's. This is a bit more detailed and has 4 targets, from 0 to 3 with 0 representing No Dementia and 3 represents Severe Dementia with different ranges as follows:
0: MMSE greater than 24 is No Dementia
1: MMSE between 20 and 24 is Mild Dementia
2: MMSE between 13 and 20 is Moderate Dementia
3: MMSE lower than 13 is Severe Dementia.